+++ grails-app/domain/org/atrocitywatch/Location.groovy
@@ -3,11 +3,12 @@ package org.atrocitywatch
 class Location {
 	
 	String name
-	long lat
-	long lon
+	double lat
+	double lon
 	double radius
 
     static constraints = {
+		
     }
 	static mapping = {
 		table '`location`'
